Quote temp dir paths in unsafe options tests
This supports temporary directory paths that require quoting when
used to build a POSIX shell command (as is done on both Unix-like
and Windows systems in this context) in the tests of the unsafe
options --upload-pack, --receive-pack, and synonyms.

Previously, test payloads in the form of touch commands were used
without quoting, which often broken on Windows, where the temporary
directory path contains backslashes, since the command is processed
as a POSIX shell command. Although far less likely, this could also
have broken on some Unix-like systems, since in principle the path
of the temporary directory prefix could be something requiring
quoting (rather than the usual /tmp).