Make real symlink test portable
This makes the test of real symlinks portable to Windows systems
where the account running the tests is able to create symlinks, by:

- Producing example nonexistent, regular-file, and directory paths
  (still absolute, as before) in a portable way. The approach taken
  is to create a temporary directory as the parent for them. Since
  in rare cases the temporary directory may be located somewhere
  whose path is not representable as ASCII, this uses the
  filesystem encoding to decode it instead. (This is the encoding
  that the code under test uses to convert the path to bytes to
  write into the object database, in IndexFile._store_path.)

- Passing a target_is_directory argument when creating the symlink.
  This is ignored on Unix-like systems but needed on Windows, where
  file and directory symlinks are distinguished by metadata in the
  symlink itself that affect their behavior.

This also does some refactoring, mainly the switch to pathlib.Path,
to limit how much more complicated the test code gets with these
changes.

Note that this commit does not fix adding real symlinks to the
index with IndexFile.add, so the test still fails for now.