Revisit installing pip in venv in separate step
The previous commit pinned the python39-pip package, but it may
still be upgraded when installed automatically in the
`python -m venv` step, since that step runs something like
`python3.9 -Im ensurepip --upgrade --default-pip`.

This replays 4605dd6 to install pip in a separate step. This
differs from the solution in #2007 in that this does not use the
bootstrap script.