Use python39-pip-wheel Cygwin package on CI
Instead of downloading and running the boostrap script.

Various Cygwin packages have been tried here for #2004, but I don't
recall having explicitly specified `python39-pip-wheel`. On the
Cygwin mailing list, Daniel Abrahamsson reported a problem that
seems to be the same as encountered here, and noted that installing
`python-pip-wheel` fixed it. So maybe that will help here. See:

https://cygwin.com/pipermail/cygwin/2025-March/257583.html

This commit also delists the `python39-virtualenv` pacakge, which
it looks like we have not been using (anytime recently) because it
actually provides the `virtualenv` module and command, while we are
using `venv` from the Python standard library, including on Cygwin.