Really try `python-pip-wheel`
In https://cygwin.com/pipermail/cygwin/2025-March/257583.html, the
package reported to fix the problem (that might be the same as this
problem) was `python-pip-wheel`. But the previous commit attempted
to fix this with `python39-pip-wheel`. This tries something that
better resembles what worked there, by listing `python39-pip` and
`python-pip-wheel` (rather than `python39-pip-wheel`).